+++ examples-pydantic-v1/movie_domain_pydantic_v1/client/data_classes/_movie.py
@@ -17,8 +17,8 @@ __all__ = ["Movie", "MovieApply", "MovieList", "MovieApplyList"]
 
 class Movie(DomainModel):
     space: ClassVar[str] = "IntegrationTestsImmutable"
-    actors: list[str] = []
-    directors: list[str] = []
+    actors: Optional[list[str]] = None
+    directors: Optional[list[str]] = None
     meta: Optional[dict] = None
     rating: Optional[str] = None
     release_year: Optional[int] = Field(None, alias="releaseYear")
@@ -40,8 +40,8 @@ class Movie(DomainModel):
 
 class MovieApply(DomainModelApply):
     space: ClassVar[str] = "IntegrationTestsImmutable"
-    actors: Union[list[ActorApply], list[str]] = Field(default_factory=list, repr=False)
-    directors: Union[list[DirectorApply], list[str]] = Field(default_factory=list, repr=False)
+    actors: Union[list[ActorApply], list[str], None] = Field(default=None, repr=False)
+    directors: Union[list[DirectorApply], list[str], None] = Field(default=None, repr=False)
     meta: Optional[dict] = None
     rating: Union[RatingApply, str, None] = Field(None, repr=False)
     release_year: Optional[int] = None
@@ -87,7 +87,7 @@ class MovieApply(DomainModelApply):
         edges = []
         cache.add(self.external_id)
 
-        for actor in self.actors:
+        for actor in self.actors or []:
             edge = self._create_actor_edge(actor)
             if edge.external_id not in cache:
                 edges.append(edge)
@@ -98,7 +98,7 @@ class MovieApply(DomainModelApply):
                 nodes.extend(instances.nodes)
                 edges.extend(instances.edges)
 
-        for director in self.directors:
+        for director in self.directors or []:
             edge = self._create_director_edge(director)
             if edge.external_id not in cache:
                 edges.append(edge)