+++ examples-pydantic-v1/shop_pydantic_v1/client/data_classes/_command_config.py
@@ -11,7 +11,7 @@ __all__ = ["CommandConfig", "CommandConfigApply", "CommandConfigList", "CommandC
 
 class CommandConfig(DomainModel):
     space: ClassVar[str] = "IntegrationTestsImmutable"
-    configs: list[str] = []
+    configs: Union[list[str], None] = None
     source: Optional[str] = None
 
     def as_apply(self) -> CommandConfigApply: