+++ examples/movie_domain/client/_api_client.py
@@ -22,7 +22,7 @@ class MovieClient:
     MovieClient
 
     Generated with:
-        pygen = 0.20.5
+        pygen = 0.21.0
         cognite-sdk = 6.25.3
         pydantic = 2.3.0