+++ examples/scenario_instance/client/data_classes/_core.py
@@ -37,14 +37,6 @@ class DomainModel(DomainModelCore):
         # can happen in there is a field named 'version' in the DominModel.
         return cls(**{**data, **unpack_properties(node.properties)})  # type: ignore[arg-type]
 
-    @classmethod
-    def one_to_many_fields(cls) -> list[str]:
-        return [
-            field_name
-            for field_name, field in cls.model_fields.items()
-            if isinstance(field.annotation, types.GenericAlias)
-        ]
-
 
 T_TypeNode = TypeVar("T_TypeNode", bound=DomainModel)