+++ examples/shop/client/_api_client.py
@@ -14,7 +14,7 @@ class ShopClient:
     ShopClient
 
     Generated with:
-        pygen = 0.20.5
+        pygen = 0.21.0
         cognite-sdk = 6.25.3
         pydantic = 2.3.0